+++ CHANGELOG.md
@@ -100,6 +100,8 @@ Shipping methods can be removed by the user after it has been assigned to a chec
   - Called when `confirmEmailChange` mutation is triggered.
 - Add `ACCOUNT_SET_PASSWORD_REQUESTED` webhook - #13486, by @Air-t
   - Called after `requestPasswordReset` or `customerCreate` mutation.
+- Add `STAFF_SET_PASSWORD_REQUESTED` webhook - #13486, by @Air-t
+  - Called after `requestPasswordReset` or `customerCreate` mutation for staff users.
 
 ### Other changes
 - Add possibility to log without confirming email - #13059 by @kadewu